Architectural changes to better separate sections
