+++ src/components/LinkLogo.tsx
@@ -1,6 +1,6 @@
 import { IconDefinition, SizeProp } from '@fortawesome/fontawesome-svg-core';
 import { useState } from 'react';
-import { LinkLogoLabel, LinkLogoWrapper } from '../styles/styledComponents';
+import { LinkLogoLabel, LinkLogoWrapper } from '../shared/styledComponents';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
 interface LinkLogoProps {
