+++ src/shared/globals.css
@@ -26,7 +26,7 @@
 }
 
 body {
-  margin: 2rem 0 0 0;
+  margin: 0; /* TODO: Set top margin to 2rem when Navbar enabled */
   display: flex;
   place-items: center;
   min-width: 320px;
